requirements.txt
cycler==0.10.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.0.3
nibabel==3.0.1
numpy==1.18.1
pandas==0.24.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
